+++ lib/components/features/event-manage/ticket/AdditionalTicketsModal.tsx
@@ -2,9 +2,9 @@ import { useState } from "react";
 
 import { Button, Input, LabeledInput, modal, ModalContent, Toggle } from "$lib/components/core";
 
-export function AdditionalTicketsModal({ ticketLimitPer, onChange }: { ticketLimitPer: number | undefined; onChange: (limit: number | undefined) => void }) {
+export function AdditionalTicketsModal({ ticketLimitPer, onChange }: { ticketLimitPer: number | null; onChange: (limit: number | null) => void }) {
   const [limit, setLimit] = useState(ticketLimitPer);
-  const [allowAdditionalTickets, setAllowAdditionalTickets] = useState(!!(ticketLimitPer === undefined || ticketLimitPer && ticketLimitPer > 1));
+  const [allowAdditionalTickets, setAllowAdditionalTickets] = useState(!!(ticketLimitPer === null || ticketLimitPer && ticketLimitPer > 1));
 
   const handleConfirm = async () => {
     onChange(limit);
@@ -37,16 +37,16 @@ export function AdditionalTicketsModal({ ticketLimitPer, onChange }: { ticketLim
             <div className="relative">
               <Input
                 type="number"
-                value={limit === undefined ? '' : limit}
+                value={limit === null ? '' : limit}
                 onChange={(e) => {
                   const val = e.target.value;
-                  setLimit(val === '' ? undefined : e.target.valueAsNumber);
+                  setLimit(val === '' ? null : e.target.valueAsNumber);
                 }}
                 variant="outlined"
                 className="[-moz-appearance:textfield] [&::-webkit-inner-spin-button]:appearance-none"
                 placeholder="Unlimited"
               />
-              <i className="icon-cancel size-5 text-quaternary cursor-pointer absolute right-3.5 top-1/2 -translate-y-1/2" onClick={() => setLimit(undefined)} />
+              <i className="icon-cancel size-5 text-quaternary cursor-pointer absolute right-3.5 top-1/2 -translate-y-1/2" onClick={() => setLimit(null)} />
             </div>
           </LabeledInput>
           )