+++ lib/components/features/event-registration/modals/RegistrationModal.tsx
@@ -24,6 +24,7 @@ import { useRedeemTickets } from "../hooks/useRedeemTickets";
 import { BuyerInfoForm } from "../forms/BuyerInfoForm";
 import { ApplicationForm } from "../forms/ApplicationForm";
 import { UserForm } from "../forms/UserInfoForm";
+import { SelfVerificationForm } from "../forms/SelfVerificationForm";
 import { CardPayment } from "../payments/CardPayment";
 import { SubmitForm } from "../SubmitForm";
 import { OrderSummary } from "../OrderSummary";
@@ -101,6 +102,9 @@ export function RegistrationModal() {
           }
           {
             !!requiredProfileFields?.length && <UserForm />
+          }
+          {
+            !!(event.self_verification?.enabled && event.self_verification?.config) && <SelfVerificationForm config={event.self_verification.config} />
           }
           {
             !!event.application_questions?.length && <ApplicationForm />